// src/schema.ts
export interface IRoleAssignment {
    Principal: string;
    RoleDefinition: string;
}

export interface IListSecurity {
    BreakRoleInheritance?: boolean;
    CopyRoleAssignments?: boolean;
    ClearSubscopes?: boolean;
    RoleAssignments?: IRoleAssignment[];
}

export interface IList {
    Title: string;
    Description: string;
    Template: number;
    ContentTypesEnabled: boolean;
    RemoveExistingContentTypes?: boolean;
    ContentTypeBindings?: IContentTypeBinding[];
    Fields?: string[];
    FieldRefs?: IListInstanceFieldRef[];
    Views?: IListView[];
    Security?: IListSecurity;

    AdditionalSettings?: {
        DefaultContentApprovalWorkflowId?: string;
        DefaultDisplayFormUrl?: string;
        DefaultEditFormUrl?: string;
        DefaultNewFormUrl?: string;
        Description?: string;
        Direction?: string;
        DocumentTemplateUrl?: string;
        /**
         * Reader = 0; Author = 1; Approver = 2.
         */
        DraftVersionVisibility?: number;
        EnableAttachments?: boolean;
        EnableFolderCreation?: boolean;
        EnableMinorVersions?: boolean;
        EnableModeration?: boolean;
        EnableVersioning?: boolean;
        ForceCheckout?: boolean;
        Hidden?: boolean;
        IrmEnabled?: boolean;
        IrmExpire?: boolean;
        IrmReject?: boolean;
        IsApplicationList?: boolean;
        NoCrawl?: boolean;
        OnQuickLaunch?: boolean;
        Title?: string;
        ValidationFormula?: string;
        ValidationMessage?: string;

        [key: string]: string | boolean | number;
    };
}

export interface IListInstanceFieldRef {
    ID: string;
    DisplayName?: string;
    Required?: boolean;
    Hidden?: boolean;
}

export interface IContentTypeBinding {
    ContentTypeID: string;
    Name?: string;
}

export interface IListView {
    Title: string;
    PersonalView?: boolean;
    ViewFields?: string[];
    AdditionalSettings?: {
        ViewQuery?: string;
        RowLimit?: number;
        Paged?: boolean;
        Hidden?: boolean;
        Scope?: 0 | 1;
    };
}